requirements.txt
Django==4.2
psycopg2_binary==2.9.6
djangorestframework==3.14.0
django_cors_headers==3.14.0
PyJWT==2.6.0
django_debug_toolbar==4.0.0
django_extensions==3.2.1
drf_yasg==1.21.5
factory_boy==3.2.1